+++ RoydeWildt/QLlib/src/main/java/org/uva/sea/ql/ast/visitor/interfaces/AtomVisitor.java
@@ -1,7 +1,7 @@
 package org.uva.sea.ql.ast.visitor.interfaces;
 
 import org.uva.sea.ql.ast.tree.atom.val.Bool;
-import org.uva.sea.ql.ast.tree.atom.val.Float;
+import org.uva.sea.ql.ast.tree.atom.val.Double;
 import org.uva.sea.ql.ast.tree.atom.val.Int;
 import org.uva.sea.ql.ast.tree.atom.val.Str;
 import org.uva.sea.ql.ast.tree.atom.var.Var;
@@ -14,7 +14,7 @@ public interface AtomVisitor<ATOM,CONTEXT> {
 
     ATOM visit(Bool atom, CONTEXT context);
     ATOM visit(Int atom, CONTEXT context);
-    ATOM visit(Float atom, CONTEXT context);
+    ATOM visit(Double atom, CONTEXT context);
     ATOM visit(Str atom, CONTEXT context);
     ATOM visit(Var atom, CONTEXT context);
