+++ RoydeWildt/QLlib/src/main/java/org/uva/sea/ql/parser/QLParser.java
@@ -8,24 +8,25 @@ package org.uva.sea.ql.parser;
     import org.uva.sea.ql.ast.tree.expr.Expr;
     import org.uva.sea.ql.ast.tree.expr.binary.*;
     import org.uva.sea.ql.ast.tree.expr.unary.*;
-    import org.uva.sea.ql.ast.tree.type.Boolean;
+	import org.uva.sea.ql.ast.tree.stat.block.If;
+	import org.uva.sea.ql.ast.tree.stat.block.IfElse;
+	import org.uva.sea.ql.ast.tree.stat.decl.Question;
+	import org.uva.sea.ql.ast.tree.type.Boolean;
     import org.uva.sea.ql.ast.tree.type.Money;
     import org.uva.sea.ql.ast.tree.type.Number;
     import org.uva.sea.ql.ast.tree.type.Text;
     import org.uva.sea.ql.ast.tree.type.Type;
     import org.uva.sea.ql.ast.tree.atom.var.*;
     import org.uva.sea.ql.ast.tree.atom.val.*;
-    import org.uva.sea.ql.ast.tree.atom.val.Float;
+    import org.uva.sea.ql.ast.tree.atom.val.Double;
     import org.uva.sea.ql.ast.tree.atom.val.Int;
 
 import org.antlr.v4.runtime.atn.*;
 import org.antlr.v4.runtime.dfa.DFA;
 import org.antlr.v4.runtime.*;
-import org.antlr.v4.runtime.misc.*;
-import org.antlr.v4.runtime.tree.*;
-import java.util.List;
-import java.util.Iterator;
-import java.util.ArrayList;
+	import org.antlr.v4.runtime.tree.*;
+
+	import java.util.ArrayList;
 
 @SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast"})
 public class QLParser extends Parser {
@@ -975,7 +976,7 @@ public class QLParser extends Parser {
 				setState(160);
 				((NumContext)_localctx).value = match(FLOAT);
 
-				            ((NumContext)_localctx).result =  new Float(((NumContext)_localctx).value, (((NumContext)_localctx).value!=null?((NumContext)_localctx).value.getText():null));
+				            ((NumContext)_localctx).result =  new Double(((NumContext)_localctx).value, (((NumContext)_localctx).value!=null?((NumContext)_localctx).value.getText():null));
 				        
 				}
 				break;
